refactor(sandbox): split MacOSSandbox into its own module & move factory to config
- Extract MacOSSandbox (+ _SEATBELT_VIOLATION_RE) from local_sandbox.py
  into a new macos_sandbox.py, matching the per-platform file convention
  already used by bubblewrap/linux/windows backends.
- Move create_sandbox() factory from local_sandbox.py into config.py so
  the type/probe/factory layer lives together; all backend branches stay
  lazily-imported to avoid circular imports (backends import .config).
- local_sandbox.py now hosts only the LocalSandbox ABC + NoneSandbox;
  drop now-unused re/Any/SandboxMode imports.
- governance: log the compiled sandbox backend (sandbox=<mode>) on each
  governance decision so the actual executor (bubblewrap/landlock/...)
  is visible in qwenpaw.log, not just action=sandbox_fallback.
- policy: allow /tmp/** by default (scratch dir, no prompt).
